updated datasource and dataset for loader
