+++ configs/selfsup/moco/r50_v2_sen12ms_in_basetrain_20ep.py
@@ -87,4 +87,4 @@ optimizer = dict(type='SGD', lr=0.03, weight_decay=0.0001, momentum=0.9)
 lr_config = dict(policy='CosineAnnealing', min_lr=0.)
 checkpoint_config = dict(interval=20)
 # runtime settings
-total_epochs = 1
+total_epochs = 2
